latest

+++ controller/clinicsController.php
@@ -3,172 +3,191 @@
 Class clinicsController Extends baseController {
 
     public function index() {
-        
+        header("Location: /index");
     }
 
     public function add() {
         session_start();
         if (Login::get_instance()->check_login() == 'valid') {
+            $user_data = Register::get_instance()->get_current_user();
+            if ($user_data['status'] != 10) {
+                header("Location: /index");
+                exit();
+            }
             $this->registry->template->title = 'Clinics | Add';
             $this->registry->template->button = 'اضافة العيادة';
             $this->registry->template->show('add_clinic');
         } else {
-            header("Location: /pets/");
-        }
-    }
-
-    public function views() {
-        session_start();
-        if (Login::get_instance()->check_login() == 'valid') {
-            $this->registry->template->title = 'Clinics | Add';
-            $this->registry->template->button = 'اضافة العيادة';
-            $this->registry->template->show('view_clinic');
-        } else {
-            header("Location: /pets/");
+            header("Location: /index");
         }
     }
 
     public function edit() {
         if (Login::get_instance()->check_login() == 'valid') {
             if ($_GET['id'] && is_numeric($_GET['id'])) {
-                $ad = ads::load_ads(5, array('id' => intval($_GET['id'])));
+                $clinic = ads::load_clinics(array('id' => intval($_GET['id'])));
                 $user_data = Register::get_instance()->get_current_user();
-                if (($user_data['id'] != $ad[0]['user_id']) && $user_data['status'] != 10) {
-                    header("Location: /pets/");
+                if (($user_data['id'] != $clinic[0]['user_id']) && $user_data['status'] != 10) {
+                    header("Location: /index");
+                    exit();
                 }
-                $this->registry->template->ad = $ad[0];
+                $this->registry->template->clinic = $clinic[0];
                 $this->registry->template->button = 'تعديل العيادة';
-                $this->registry->template->title = 'Pets | Edit | ' . $ad[0]['title'];
-                $this->registry->template->show('add_ad');
+                $this->registry->template->title = 'Pets | Edit | ' . $clinic[0]['name'];
+                $this->registry->template->show('add_clinic');
             } else {
-                header("Location: /pets/ads");
+                header("Location: /index");
             }
         } else {
-            header("Location: /pets/");
+            header("Location: /index");
         }
     }
 
     public function view() {
         if ($_GET['id'] && is_numeric($_GET['id'])) {
-            $ad = ads::load_ads(5, array('id' => intval($_GET['id'])));
-            $this->registry->template->activate = 0;
-            $this->registry->template->ad = $ad[0];
-            $user = Register::get_instance()->get_user(array('id' => $ad[0]['user_id']));
-            $this->registry->template->phone = $user[0]['phone'];
-            $this->registry->template->title = 'Pets | ' . $ad[0]['title'];
-            $this->registry->template->show('view_ad');
+            $clinic = ads::load_clinics(array('id' => intval($_GET['id'])));
+            $this->registry->template->clinic = $clinic[0];
+            $this->registry->template->title = 'Pets | ' . $clinic[0]['name'];
+            $this->registry->template->show('view_clinic');
         } else {
-            header("Location: /pets/");
+            header("Location: /index");
         }
     }
 
-    public function processad() {
+    public function processclinic() {
         if (Login::get_instance()->check_login() == 'valid' && $_POST) {
+            $user_data = Register::get_instance()->get_current_user();
+            if ($user_data['status'] != 10) {
+                exit();
+            }
             $errors['rt'] = array();
             $errors['md'] = array();
             $errors['lt'] = array();
 
-            $ad_data = array();
+            $clinic_data = array();
             $user_data = Register::get_instance()->get_current_user();
 
-            if (!empty($_POST['type']))
-                $ad_data['type'] = intval($_POST['type']);
+            if (!empty($_POST['doc_name']) && trim($_POST['doc_name']) != '')
+                $clinic_data['doc_name'] = addslashes($_POST['doc_name']);
             else
-                $errors['rt'][] = 'يجب اختيار نوع الاعلان';
+                $errors['rt'][] = 'يجب ادخال اسم الدكتور';
 
-            if (!empty($_POST['pet']))
-                $ad_data['pet_id'] = intval($_POST['pet']);
+            if (!empty($_POST['phone1']) && is_numeric($_POST['phone1']))
+                $clinic_data['phone1'] = $_POST['phone1'];
             else
-                $errors['rt'][] = 'يجب اختيار نوع الحيوان';
+                $errors['rt'][] = 'يجب ادخال رقم التليفون  1بشكل صحيح';
 
-            if (!empty($_POST['title']) && trim($_POST['title']) != '')
-                $ad_data['title'] = addslashes($_POST['title']);
+            if (!empty($_POST['phone2'])) {
+                if (!is_numeric($_POST['phone2']))
+                    $errors['rt'][] = 'يجب ادخال رقم التليفون 2 بشكل صحيح';
+                else
+                    $clinic_data['phone2'] = $_POST['phone2'];
+            }
+
+            if (!empty($_POST['phone3'])) {
+                if (!is_numeric($_POST['phone3']))
+                    $errors['rt'][] = 'يجب ادخال رقم التليفون 3 بشكل صحيح';
+                else
+                    $clinic_data['phone3'] = $_POST['phone3'];
+            }
+
+            if (!empty($_POST['name']) && trim($_POST['name']) != '')
+                $clinic_data['name'] = addslashes($_POST['name']);
             else
-                $errors['md'][] = 'يجب ادخال عنوان للاعلان';
+                $errors['md'][] = 'يجب ادخال اسم العيادة';
+
+
 
             if (!empty($_POST['desc']) && trim($_POST['desc']) != '')
-                $ad_data['desc'] = addslashes($_POST['desc']);
+                $clinic_data['desc'] = addslashes($_POST['desc']);
             else
-                $errors['md'][] = 'يجب ادخال معلومات الاعلان';
+                $errors['md'][] = 'يجب ادخال مواعيد عمل العيادة';
+
+            if (!empty($_POST['address']) && trim($_POST['address']) != '')
+                $clinic_data['address'] = addslashes($_POST['address']);
+            else
+                $errors['lt'][] = 'يجب ادخال عنوان العيادة';
 
             if (!empty($_POST['country']))
-                $ad_data['country'] = intval($_POST['country']);
+                $clinic_data['country'] = intval($_POST['country']);
             else {
                 $errors['lt'][] = 'يجب اختيار الدولة';
                 goto a;
             }
             if (!empty($_POST['city']))
-                $ad_data['city'] = intval($_POST['city']);
+                $clinic_data['city'] = intval($_POST['city']);
             else {
                 $errors['lt'][] = 'يجب اختيار المدينة';
                 goto a;
             }
 
             if (!empty($_POST['region']))
-                $ad_data['region'] = intval($_POST['region']);
+                $clinic_data['region'] = intval($_POST['region']);
             else
                 $errors['lt'] = 'يجب اختيار المنطقة';
 
-            $ad_data['cat_id'] = intval($_POST['cat']);
-            $ad_data['user_id'] = $user_data['id'];
-            $ad_data['time_added'] = time();
-            $ad_data['date_added'] = TimeTools::get_time_id(date('Y-m-d'));
-
+            $clinic_data['user_id'] = $user_data['id'];
+            $clinic_data['time_added'] = time();
+            $clinic_data['date_added'] = TimeTools::get_time_id(date('Y-m-d'));
+            $clinic_data['status'] = 1;
             a:
             if ($_POST['id'] != 'null') {
-                $ad_data['status'] = 0;
                 $op = 'update';
                 if (is_numeric($_POST['id']))
-                    $ad_data['id'] = intval($_POST['id']);
+                    $clinic_data['id'] = intval($_POST['id']);
                 else
-                    $errors['rt'][] = 'خطأ فى اضافة الاعلان';
-
-                $img_check = $this->check_img($ad_data['user_id'], $ad_data['id']);
+                    $errors['rt'][] = 'خطأ فى اضافة العيادة';
                 $type = 'edit';
-            }else {
+            } else {
                 $op = 'insert';
-                $img_check = $this->check_img($ad_data['user_id'], 'null');
                 $type = 'add';
             }
-            if (($img_check !== true) && empty($errors['rt']))
-                $errors['rt'][] = $img_check;
-
 
             if (empty($errors['rt']) && empty($errors['md']) && empty($errors['lt'])) {
-                if ($op == 'insert')
-                    $ad_id = Operations::get_instance()->init($ad_data, 'ads');
-                else {
-                    $ad_id = $ad_data['id'];
-                    Operations::get_instance()->init($ad_data, 'ads', 'update');
+                if ($op == 'insert') {
+                    $clinic_id = Operations::get_instance()->init($clinic_data, 'clinics');
+                    Operations::get_instance()->init(
+                            array(
+                        'img_name' => '.jpeg',
+                        'clinic_id' => $clinic_id,
+                        'time_added' => time(),
+                        'date_added' => TimeTools::get_time_id(date('Y-m-d')),
+                            ), 'clinics_img');
+                } else {
+                    $clinic_id = $clinic_data['id'];
+                    Operations::get_instance()->init($clinic_data, 'clinics', 'update');
                 }
 
-                $this->procces_img($ad_data['user_id'], $ad_id);
-                echo json_encode(array('operation' => 1, 'type' => $type));
+                $this->img_upload($clinic_id);
+                echo json_encode(array('operation' => 1, 'type' => $type, 'id' => $clinic_id));
             } else {
                 echo json_encode(array('operation' => 2, 'errors' => $errors));
             }
         }
     }
 
-    private function img_upload() {
+    private function img_upload($clinic_id) {
         if (Login::get_instance()->check_login() == 'valid') {
+            $user_data = Register::get_instance()->get_current_user();
+            if ($user_data['status'] != 10) {
+                exit();
+            }
             $foo = new Upload($_FILES['clinic_img']);
             if ($foo->uploaded) {
-                $this->del();
+                $this->del($clinic_id);
                 $ds = DIRECTORY_SEPARATOR;
-                $user_data = Register::get_instance()->get_current_user();
                 $storeFolder = '..' . $ds . 'views' . $ds . 'clinics_img';
                 $targetPath = dirname(__FILE__) . $ds . $storeFolder . $ds;
-                $targetName = $user_data['id'] . '_' . md5(rand(1, 5000000000));
+                $targetName = $clinic_id . '_' . md5(rand(1, 5000000000));
                 while (file_exists($targetName)) {
-                    $targetName = $user_data['id'] . '_' . md5(rand(1, 5000000000));
+                    $targetName = $clinic_id . '_' . md5(rand(1, 5000000000));
                 }
                 $targetFile = $targetPath . $targetName . '.jpeg';
                 $foo->file_new_name_body = $targetName;
                 $foo->image_resize = true;
                 $foo->image_convert = 'jpeg';
-                $foo->image_x = 1000;
-                $foo->image_y = 1000;
+                $foo->image_x = 1024;
+                $foo->image_y = 768;
                 $foo->image_ratio_crop = false;
                 //$foo->image_ratio_y = true;
                 $foo->Process($targetPath);
@@ -177,22 +196,25 @@ Class clinicsController Extends baseController {
                     Operations::get_instance()->init(
                             array(
                         'img_name' => $targetName . '.jpeg',
-                        'user_id' => $user_data['id'],
+                        'clinic_id' => $clinic_id,
                         'time_added' => time(),
                         'date_added' => TimeTools::get_time_id(date('Y-m-d')),
-                            ), 'users_img', 'update');
+                            ), 'clinics_img', 'update');
                 }
             }
         }
     }
 
-    public function del() {
+    public function del($clinic_id) {
         if (Login::get_instance()->check_login() == 'valid') {
-            $ds = DIRECTORY_SEPARATOR;
             $user_data = Register::get_instance()->get_current_user();
-            $storeFolder = '..' . $ds . 'views' . $ds . 'users_img';
+            if ($user_data['status'] != 10) {
+                exit();
+            }
+            $ds = DIRECTORY_SEPARATOR;
+            $storeFolder = '..' . $ds . 'views' . $ds . 'clinics_img';
             $targetPath = dirname(__FILE__) . $ds . $storeFolder . $ds;
-            $images = glob($targetPath . $user_data['id'] . "_*.jpeg");
+            $images = glob($targetPath . $clinic_id . "_*.jpeg");
             foreach ($images as $image) {
                 unlink($image);
             }
